Modularize routes.ts: extract flows + env-validate
Adopting the 250-350 line per file rule. routes.ts is the worst
offender at 2,101 lines (6x the limit); this is wave 1 of breaking
it up by domain. Each extracted module has a single objective and
fits cleanly under the cap.

server/routes-modules/flows.ts  (148 lines)
  All /api/admin/flows/* CRUD + publish/archive/duplicate, plus
  user-facing /api/flows + run lifecycle (start, list, detail,
  approve, reject, resume). Route-order requirement (/api/flows/runs
  before /api/flows/:id) documented inside the module so future
  edits can't accidentally re-introduce the shadowing bug.
  Exposes: registerFlowRoutes(app).

server/services/EnvValidator.ts  (327 lines, pure)
  Pure validation of process.env. No HTTP — returns
  { ok, summary, checks }. Helpers split by concern:
    - checkUrl, pushProviderChecks, pushSessionSecretCheck,
      pushDatabaseCheck, pushAdminAuthChecks
  Top-level validateEnv() composes them. Now unit-testable and
  callable from CLI smoke tests / CI gates.

server/routes-modules/env-validate.ts  (11 lines)
  Thin route wrapper: GET /api/admin/env-validate calls validateEnv()
  and returns the result. Exposes: registerEnvValidateRoute(app).

server/routes.ts  (2,101 → 1,697 lines)
  Removed the 124-line inline flow registration + 283-line inline
  env-validate handler. Replaced with two import + register calls.
  Net -404 lines from routes.ts.

Next waves (not in this PR): project routes, conversation/messages,
auth, knowledge, ruleset, voice, security — same pattern, each into
its own routes-modules/<topic>.ts file.
