+++ server/routes.ts
@@ -21,13 +21,9 @@ import { KnowledgeService } from "./services/KnowledgeService";
 import { getFirewallIntegrationStatus } from "./services/FirewallIntegrationService";
 import { checkGitHubIntegrationStatus, getGitHubRepoReadout } from "./services/GitHubIntegrationService";
 import { getRecentRuntimeEvents, logRuntimeEvent } from "./services/RuntimeLogger";
-import { FlowStore } from "./services/FlowStore";
 import { buildZedAdminContext } from "./services/ZedContextBuilder";
-import {
-  executeFlowRun,
-  approveCurrentStage,
-  rejectCurrentStage,
-} from "./services/flow/FlowExecutor";
+import { registerFlowRoutes } from "./routes-modules/flows";
+import { registerEnvValidateRoute } from "./routes-modules/env-validate";
 import { registerExecutionRoutes } from "./services/execution/registerExecutionRoutes";
 import { registerIntakeRoutes } from "./services/intake/registerIntakeRoutes";
 import fs from "fs/promises";
@@ -1439,414 +1435,14 @@ export async function registerRoutes(app: Express): Promise<Server> {
     });
   });
 
-  app.get("/api/admin/flows", isAdmin, async (req: any, res) => {
-    const includeArchived = String(req.query.includeArchived || "") === "true";
-    const flows = await FlowStore.listDefinitions({ includeArchived });
-    res.json({ flows });
-  });
+  // ── Flows (admin CRUD + user-facing + run lifecycle) ──────────────
+  // Extracted to server/routes-modules/flows.ts. Route-order requirement
+  // (/api/flows/runs before /api/flows/:id) is preserved inside the module.
+  registerFlowRoutes(app);
 
-  app.get("/api/admin/flows/:id", isAdmin, async (req, res) => {
-    const flow = await FlowStore.getDefinition(req.params.id);
-    if (!flow) return res.status(404).json({ error: "Not found" });
-    res.json(flow);
-  });
-
-  app.post("/api/admin/flows", isAdmin, async (req, res) => {
-    try {
-      const flow = await FlowStore.createDefinition(req.body);
-      res.json(flow);
-    } catch (err: any) {
-      res.status(400).json({ error: err?.message || "Failed to create flow" });
-    }
-  });
-
-  app.put("/api/admin/flows/:id", isAdmin, async (req, res) => {
-    const flow = await FlowStore.updateDefinition(req.params.id, req.body);
-    if (!flow) return res.status(404).json({ error: "Not found" });
-    res.json(flow);
-  });
-
-  app.post("/api/admin/flows/:id/publish", isAdmin, async (req, res) => {
-    const flow = await FlowStore.publishDefinition(req.params.id);
-    if (!flow) return res.status(404).json({ error: "Not found" });
-    res.json(flow);
-  });
-
-  app.post("/api/admin/flows/:id/archive", isAdmin, async (req, res) => {
-    const flow = await FlowStore.archiveDefinition(req.params.id);
-    if (!flow) return res.status(404).json({ error: "Not found" });
-    res.json(flow);
-  });
-
-  app.post("/api/admin/flows/:id/duplicate", isAdmin, async (req, res) => {
-    const flow = await FlowStore.duplicateDefinition(req.params.id);
-    if (!flow) return res.status(404).json({ error: "Not found" });
-    res.json(flow);
-  });
-
-  app.get("/api/flows", isAuthenticated, async (_req, res) => {
-    const flows = await FlowStore.listPublished();
-    res.json({
-      flows: flows.map((f) => ({
-        id: f.id,
-        slug: f.slug,
-        name: f.name,
-        category: f.category,
-        userFacingLabel: f.userFacingLabel,
-        userFacingBlurb: f.userFacingBlurb,
-        icon: f.icon,
-        stageCount: f.stages.length,
-        agents: f.agents,
-      })),
-    });
-  });
-
-  // ⚠️ /api/flows/runs and /api/flows/runs/:runId MUST be registered
-  // before /api/flows/:id, otherwise Express matches "runs" as the :id
-  // parameter and the runs list 404s.
-  app.get("/api/flows/runs", isAuthenticated, async (req: any, res) => {
-    const runs = await FlowStore.listRuns({
-      userId: req.user?.claims?.sub,
-      limit: 50,
-    });
-    res.json({ runs });
-  });
-
-  app.get("/api/flows/runs/:runId", isAuthenticated, async (req, res) => {
-    const run = await FlowStore.getRun(req.params.runId);
-    if (!run) return res.status(404).json({ error: "Not found" });
-    res.json(run);
-  });
-
-  app.get("/api/flows/:id", isAuthenticated, async (req, res) => {
-    const flow = await FlowStore.getDefinition(req.params.id);
-    if (!flow || flow.status !== "published") {
-      return res.status(404).json({ error: "Not found" });
-    }
-    res.json(flow);
-  });
-
-  app.post("/api/flows/:id/run", isAuthenticated, async (req: any, res) => {
-    const flow = await FlowStore.getDefinition(req.params.id);
-    if (!flow || flow.status !== "published") {
-      return res.status(404).json({ error: "Not found" });
-    }
-    const run = await FlowStore.startRun({
-      flow,
-      userId: req.user?.claims?.sub || "anonymous",
-      conversationId: req.body?.conversationId,
-      context: req.body?.context,
-    });
-    // Kick off async execution. Don't block the response — the UI will
-    // poll the run detail endpoint to observe progress.
-    void executeFlowRun(run.id);
-    res.json(run);
-  });
-
-  app.post("/api/flows/runs/:runId/approve", isAuthenticated, async (req, res) => {
-    const run = await approveCurrentStage(req.params.runId, req.body?.note);
-    if (!run) return res.status(404).json({ error: "Run not found or not pending approval" });
-    res.json(run);
-  });
-
-  app.post("/api/flows/runs/:runId/reject", isAuthenticated, async (req, res) => {
-    const run = await rejectCurrentStage(req.params.runId, req.body?.reason);
-    if (!run) return res.status(404).json({ error: "Run not found or not pending approval" });
-    res.json(run);
-  });
-
-  app.post("/api/flows/runs/:runId/resume", isAuthenticated, async (req, res) => {
-    // Manual nudge if a run got stuck (e.g. process restarted mid-execution).
-    const run = await FlowStore.getRun(req.params.runId);
-    if (!run) return res.status(404).json({ error: "Not found" });
-    void executeFlowRun(req.params.runId);
-    res.json({ ok: true, runId: req.params.runId });
-  });
-
-  app.get("/api/admin/env-validate", isAdmin, async (_req, res) => {
-    type Severity = "ok" | "warn" | "error";
-    interface EnvCheck {
-      name: string;
-      severity: Severity;
-      message: string;
-      hint?: string;
-    }
-    const checks: EnvCheck[] = [];
-    const env = process.env;
-
-    const trimmed = (k: string) => (env[k] ?? "").trim();
-    const present = (k: string) => trimmed(k).length > 0;
-
-    const provider = (env.MODEL_PROVIDER || "").trim().toLowerCase();
-    if (!provider) {
-      checks.push({
-        name: "MODEL_PROVIDER",
-        severity: "error",
-        message: "Not set. Active provider cannot be determined.",
-        hint: 'Set to one of: "openai", "claude", "ollama", "claw-temp".',
-      });
-    } else if (!["openai", "claude", "ollama", "claw-temp"].includes(provider)) {
-      checks.push({
-        name: "MODEL_PROVIDER",
-        severity: "error",
-        message: `Unknown value "${provider}".`,
-        hint: 'Must be one of: "openai", "claude", "ollama", "claw-temp".',
-      });
-    } else {
-      checks.push({
-        name: "MODEL_PROVIDER",
-        severity: "ok",
-        message: `Active provider is "${provider}".`,
-      });
-    }
-
-    function checkUrl(name: string, expectedSuffix?: string) {
-      const raw = trimmed(name);
-      if (!raw) return null as EnvCheck | null;
-      const malformedChars = /[<>"`\s]/;
-      if (malformedChars.test(raw)) {
-        return {
-          name,
-          severity: "error" as Severity,
-          message: `Contains illegal characters (one of: < > " \` whitespace). Likely a copy-paste mistake.`,
-          hint: "Remove any angle brackets / quotes; the value should be the raw URL only.",
-        };
-      }
-      let url: URL;
-      try {
-        url = new URL(raw);
-      } catch {
-        return {
-          name,
-          severity: "error" as Severity,
-          message: `"${raw.slice(0, 80)}" is not a valid URL.`,
-        };
-      }
-      if (!["http:", "https:"].includes(url.protocol)) {
-        return {
-          name,
-          severity: "error" as Severity,
-          message: `Expected http(s) but got ${url.protocol}.`,
-        };
-      }
-      if (raw.endsWith("/")) {
-        return {
-          name,
-          severity: "warn" as Severity,
-          message: "URL has a trailing slash; some gateways double-up paths.",
-          hint: "Remove the trailing slash.",
-        };
-      }
-      if (expectedSuffix && !raw.toLowerCase().endsWith(expectedSuffix.toLowerCase())) {
-        return {
-          name,
-          severity: "warn" as Severity,
-          message: `Doesn't end in "${expectedSuffix}". Most OpenAI-compatible providers expect a base URL ending there.`,
-          hint: `Try ${raw.replace(/\/+$/, "")}${expectedSuffix} unless your provider documents a different path.`,
-        };
-      }
-      return {
-        name,
-        severity: "ok" as Severity,
-        message: `${url.host}${url.pathname || ""} — looks well-formed.`,
-      };
-    }
-
-    if (provider === "openai") {
-      if (!present("OPENAI_API_KEY")) {
-        checks.push({
-          name: "OPENAI_API_KEY",
-          severity: "error",
-          message: "Required when MODEL_PROVIDER=openai but not set.",
-        });
-      } else {
-        checks.push({
-          name: "OPENAI_API_KEY",
-          severity: "ok",
-          message: `Set (length ${trimmed("OPENAI_API_KEY").length}).`,
-        });
-      }
-      const urlCheck = checkUrl("OPENAI_BASE_URL", "/v1");
-      if (urlCheck) checks.push(urlCheck);
-      else
-        checks.push({
-          name: "OPENAI_BASE_URL",
-          severity: "warn",
-          message: "Not set; falling back to https://api.openai.com/v1.",
-          hint: "Set explicitly when using a non-OpenAI gateway like Lightning AI.",
-        });
-      if (!present("OPENAI_MODEL") && !present("MODEL_NAME")) {
-        checks.push({
-          name: "OPENAI_MODEL",
-          severity: "warn",
-          message: "Not set. Default model will be used (gpt-4o-mini).",
-          hint: "Set OPENAI_MODEL to your gateway's model identifier.",
-        });
-      } else {
-        checks.push({
-          name: "OPENAI_MODEL",
-          severity: "ok",
-          message: `Default model: ${trimmed("OPENAI_MODEL") || trimmed("MODEL_NAME")}.`,
-        });
-      }
-    }
-
-    if (provider === "claude") {
-      if (!present("CLAUDE_API_KEY") && !present("ANTHROPIC_API_KEY")) {
-        checks.push({
-          name: "CLAUDE_API_KEY",
-          severity: "error",
-          message: "Required when MODEL_PROVIDER=claude but neither CLAUDE_API_KEY nor ANTHROPIC_API_KEY is set.",
-        });
-      }
-      const urlCheck = checkUrl("CLAUDE_BASE_URL");
-      if (urlCheck) checks.push(urlCheck);
-    }
-
-    if (provider === "ollama") {
-      const urlCheck = checkUrl("OLLAMA_URL");
-      if (urlCheck) checks.push(urlCheck);
-    }
-
-    const lanes = ["CHAT", "MANAGER", "OPERATIONS", "RESEARCH", "BUSINESS", "FINANCE"];
-    const overrideCount = lanes.filter((lane) => present(`MODEL_${lane}`)).length;
-    if (overrideCount > 0) {
-      checks.push({
-        name: "MODEL_<lane> overrides",
-        severity: "ok",
-        message: `${overrideCount} of ${lanes.length} lanes have explicit overrides.`,
-      });
-    }
-
-    const sessionSecret = trimmed("SESSION_SECRET");
-    if (!sessionSecret) {
-      checks.push({
-        name: "SESSION_SECRET",
-        severity: "error",
-        message: "Not set. Session cookies cannot be signed; logins will fail.",
-        hint: "Generate a 32-byte random hex string.",
-      });
-    } else if (sessionSecret.length < 24) {
-      checks.push({
-        name: "SESSION_SECRET",
-        severity: "error",
-        message: `Only ${sessionSecret.length} characters — too short to be cryptographically strong.`,
-        hint: "Use at least 32 random characters (e.g. openssl rand -hex 32).",
-      });
-    } else if (/@/.test(sessionSecret) || /^[a-zA-Z]+$/.test(sessionSecret)) {
-      checks.push({
-        name: "SESSION_SECRET",
-        severity: "error",
-        message: "Looks like an email or simple word, not a random secret.",
-        hint: "Replace with a high-entropy random string (32+ bytes).",
-      });
-    } else if (/^(password|secret|admin|test|changeme)/i.test(sessionSecret)) {
-      checks.push({
-        name: "SESSION_SECRET",
-        severity: "warn",
-        message: "Starts with a common dictionary word.",
-      });
-    } else {
-      checks.push({
-        name: "SESSION_SECRET",
-        severity: "ok",
-        message: `Set (length ${sessionSecret.length}).`,
-      });
-    }
-
-    const dbUrl = trimmed("DATABASE_URL");
-    if (!dbUrl) {
-      checks.push({
-        name: "DATABASE_URL",
-        severity: "error",
-        message: "Not set. Database access will fail.",
-      });
-    } else if (!/^postgres(ql)?:\/\//.test(dbUrl)) {
-      checks.push({
-        name: "DATABASE_URL",
-        severity: "error",
-        message: 'Does not start with "postgres://" or "postgresql://".',
-        hint: "Drizzle expects a Postgres connection string.",
-      });
-    } else {
-      try {
-        const u = new URL(dbUrl);
-        checks.push({
-          name: "DATABASE_URL",
-          severity: "ok",
-          message: `${u.hostname}${u.pathname} — well-formed Postgres URL.`,
-        });
-      } catch {
-        checks.push({
-          name: "DATABASE_URL",
-          severity: "error",
-          message: "Could not parse as a URL.",
-        });
-      }
-    }
-
-    if (!present("ZED_ADMIN_USERNAME")) {
-      checks.push({
-        name: "ZED_ADMIN_USERNAME",
-        severity: "error",
-        message: "Not set. Admin login form will reject every attempt.",
-      });
-    } else {
-      checks.push({
-        name: "ZED_ADMIN_USERNAME",
-        severity: "ok",
-        message: `Set to "${trimmed("ZED_ADMIN_USERNAME")}".`,
-      });
-    }
-
-    if (!present("ZED_ADMIN_PASSWORD") && !present("ZED_ADMIN_SECURE_PHRASE")) {
-      checks.push({
-        name: "ZED_ADMIN_PASSWORD",
-        severity: "error",
-        message:
-          "Neither ZED_ADMIN_PASSWORD nor ZED_ADMIN_SECURE_PHRASE is set. Admin login is impossible.",
-      });
-    } else if (
-      present("ZED_ADMIN_PASSWORD") &&
-      trimmed("ZED_ADMIN_PASSWORD").length < 10
-    ) {
-      checks.push({
-        name: "ZED_ADMIN_PASSWORD",
-        severity: "warn",
-        message: `Only ${trimmed("ZED_ADMIN_PASSWORD").length} characters — short.`,
-        hint: "Use 16+ characters for a public deploy.",
-      });
-    } else {
-      checks.push({
-        name: "ZED_ADMIN_PASSWORD",
-        severity: "ok",
-        message: "Set with reasonable length.",
-      });
-    }
-
-    const frontendUrlCheck = checkUrl("FRONTEND_URL");
-    if (frontendUrlCheck) checks.push(frontendUrlCheck);
-
-    if (present("BRAVE_SEARCH_API_KEY")) {
-      checks.push({
-        name: "BRAVE_SEARCH_API_KEY",
-        severity: "ok",
-        message: "Set — web search via Brave is wired.",
-      });
-    }
-
-    const summary = {
-      ok: checks.filter((c) => c.severity === "ok").length,
-      warn: checks.filter((c) => c.severity === "warn").length,
-      error: checks.filter((c) => c.severity === "error").length,
-    };
-
-    res.json({
-      ok: summary.error === 0,
-      summary,
-      checks,
-    });
-  });
+  // Env validator — pure logic in services/EnvValidator.ts, thin route
+  // wrapper in routes-modules/env-validate.ts.
+  registerEnvValidateRoute(app);
 
   app.get("/api/admin/ruleset/structured", isAdmin, async (_req, res) => {
     const files = ["personality.yaml", "security.yaml", "parameters.yaml", "access.yaml"];
